+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## 🐛 Bugfixes
 - fixed a bug with Dockerfile path resolution
-- ramping up runner conformance from 160/378 to 162/378
+- handle NetworkAccess Requirment in runner
+- ramping up runner conformance from 160/378 to 163/378
 
 # v0.5.2
 ## 🐛 Bugfixes